remove PTS URL &  generation
